deploy11

+++ database/migrations/2022_06_04_225116_add_plan_to_subscriptions.php
@@ -15,7 +15,7 @@ class AddPlanToSubscriptions extends Migration
     {
         Schema::table('subscriptions', function (Blueprint $table) {
             $table->bigInteger('property_listing_id')->after('plot_type')->unsigned();
-            $table->integer('property_listing_plan_id')->after('property_listing_id')->unsigned();
+            $table->bigInteger('property_listing_plan_id')->after('property_listing_id')->unsigned();
 
             $table->foreign('property_listing_id')->references('id')->on('property_listings');
             $table->foreign('property_listing_plan_id')->references('id')->on('property_listing_plans');
